move tests for elasticsearch pollForSnapshotCreation
